Link stuck fixed, premium download issue fixed, home page categories boxes displayed and crud completed in admin

+++ client/src/pages/Admin/Categories/index.jsx
@@ -22,6 +22,9 @@ export default function Categories() {
     const [subcategory, setSubcategory] = useState("")
     const [loading, setLoading] = useState(false)
 
+    const [selectedFile, setSelectedFile] = useState(null)
+    const [preview, setPreview] = useState(null)
+
     useEffect(() => {
         if (userData?.userID) {
             fetchCategories()
@@ -46,14 +49,32 @@ export default function Categories() {
             })
     }
 
-    const handleCreateCategory = () => {
-        const newCategory = {
-            category: createCategoryInput.toLowerCase(),
-            subcategories: []
+    const handleFileChange = e => {
+        const file = e.target.files[0]
+        if (file) {
+            const allowedTypes = ['image/jpeg', 'image/png', 'image/webp']
+            if (!allowedTypes.includes(file.type)) {
+                return window.toastify("Only PNG, JPG or WEBP files are allowed!", "error")
+            }
+            setPreview(URL.createObjectURL(file))
+            setSelectedFile(file)
         }
+    }
+
+    const handleCreateCategory = () => {
+        const formData = new FormData();
+        formData.append("image", selectedFile);
+        formData.append("category", createCategoryInput.toLowerCase());
 
         setLoading(true)
-        axios.post(`${import.meta.env.VITE_HOST}/admin/create-category`, newCategory)
+        axios.post(`${import.meta.env.VITE_HOST}/admin/create-category`,
+            formData,
+            {
+                headers: {
+                    "Content-Type": "multipart/form-data"
+                }
+            }
+        )
             .then(res => {
                 const { status, data } = res
                 if (status === 201) {
@@ -63,48 +84,76 @@ export default function Categories() {
                 }
             })
             .catch(err => {
-                console.error('Frontend POST error', err.message)
-                if (err.response.status === 400) {
-                    window.toastify(err.response.data.message, "error")
-                }
-                else {
-                    window.toastify("Something went wrong. Please try again!", "error")
-                }
+                console.error('Frontend POST error', err);
+                const message = err.response?.data?.message ||
+                    err.message ||
+                    "Something went wrong. Please try again!";
+                window.toastify(message, "error");
             })
             .finally(() => {
                 setLoading(false)
                 setCreateCategoryInput("")
+                setSelectedFile(null)
+                setPreview(null)
             })
     }
 
     const handleUpdateCategory = (categoryDetails) => {
         setOpenCatModel(true)
         setUpdatingCategory(categoryDetails)
+        setPreview(null);
+        setSelectedFile(null);
     }
 
     const handleUpdateCategoryFunction = () => {
+        if (!updatingCategory?.category?.trim()) {
+            return window.toastify("Category name is required", "error");
+        }
+
         setLoading(true)
-        axios.patch(`${import.meta.env.VITE_HOST}/admin/update-category`, updatingCategory)
+
+        const formData = new FormData();
+        formData.append("_id", updatingCategory._id);
+        formData.append("category", updatingCategory.category.toLowerCase());
+
+        if (selectedFile) {
+            formData.append("image", selectedFile);
+        }
+
+        axios.patch(`${import.meta.env.VITE_HOST}/admin/update-category`,
+            formData,
+            {
+                headers: {
+                    "Content-Type": "multipart/form-data"
+                }
+            }
+        )
             .then(res => {
                 const { status, data } = res
                 if (status === 202) {
-                    const updatedCategories = categories.map(cat => cat._id === updatingCategory._id ? { ...cat, category: updatingCategory.category } : cat)
+                    const updatedCategories = categories.map(cat => cat._id === updatingCategory._id ? { ...cat, category: updatingCategory.category, imageURL: data.updatedCategory.imageURL } : cat)
                     setCategories(updatedCategories)
                     window.toastify(data.message, "success")
                 }
             })
             .catch(err => {
-                console.error('Frontend POST error', err.message)
-                window.toastify("Something went wrong. Please try again!", "error")
+                console.error('Update error:', err);
+                const message = err.response?.data?.message || err.message || "Something went wrong. Please try again!";
+                window.toastify(message, "error");
             })
             .finally(() => {
                 setLoading(false)
                 setOpenCatModel(false)
                 setUpdatingCategory({})
+                setPreview(null)
+                setSelectedFile(null)
             })
     }
 
     const handleCategoryDelete = (delCatID) => {
+        const confirmDelete = window.confirm("Are you sure you want to delete this category?");
+        if (!confirmDelete) return
+
         setLoading(true)
         axios.delete(`${import.meta.env.VITE_HOST}/admin/delete-category/${delCatID}`)
             .then(res => {
@@ -208,47 +257,66 @@ export default function Categories() {
                 </button>
             </div>
 
-            <div className='categories-container'>
-                {
-                    categories.map((cat, i) => {
-                        return (
-                            <div key={i} className='category-box'>
-                                <div>
-                                    <p className='font-bold capitalize !text-[#333]'>{i + 1}. {cat.category}</p>
-                                    <div className="flex flex-wrap justify-start gap-2 mt-2">
-                                        {
-                                            cat.subcategories.length > 0 ?
-                                                cat.subcategories?.map((sub, j) => (
-                                                    <span
-                                                        key={j}
-                                                        className="px-3 py-1 bg-[var(--md-light)] text-[var(--primary)] !text-[12px] text-sm rounded-full capitalize"
-                                                    >
-                                                        {sub}
-                                                    </span>
-                                                ))
-                                                :
-                                                <p className='!text-[14px]'>There is no subcategory</p>
-                                        }
-                                    </div>
-                                </div>
-                                <div className='flex justify-end gap-5 mt-5'>
-                                    <CgEditExposure className='text-[16px] text-blue-500 cursor-pointer hover:text-blue-300' onClick={() => handleSubCategory(cat)} />
-                                    <MdEdit className='text-[16px] text-blue-500 cursor-pointer hover:text-blue-300' onClick={() => handleUpdateCategory(cat)} />
-                                    <CgTrashEmpty className='text-[16px] text-red-500 cursor-pointer hover:text-red-300' onClick={() => handleCategoryDelete(cat._id)} />
-                                </div>
-                            </div>
-                        )
-                    })
-                }
+            <div className='w-full overflow-x-auto mt-8 rounded-[12px]'>
+                <table className='min-w-[600px] w-full text-sm text-left bg-white'>
+                    <thead className='bg-[var(--md-light)] text-[var(--dark)] uppercase text-xs'>
+                        <tr>
+                            <th className='p-4'>Sr#</th>
+                            <th className='p-4'>Image</th>
+                            <th className='p-4'>Category</th>
+                            <th className='p-4'>Subcategories</th>
+                            <th className='p-4 text-end'>Actions</th>
+                        </tr>
+                    </thead>
+                    <tbody>
+                        {categories.length > 0 ? (
+                            categories.map((cat, i) => (
+                                <tr key={cat._id} className='hover:bg-gray-50'>
+                                    <td className='p-4'>{i + 1}</td>
+                                    <td className='p-4'><img src={`${import.meta.env.VITE_ASURA_SUBDOMAIN}${cat.imageURL}`} alt={i + 1} className='w-10 h-10 object-contain' /></td>
+                                    <td className='p-4 capitalize'>{cat.category}</td>
+                                    <td className='flex gap-1 p-4 capitalize'>{cat.subcategories.slice(0, 4).map((sub, i) => (
+                                        <span key={i} className='!text-[14px] text-[var(--dark)] bg-[var(--md-light)] rounded-full px-2'>{sub}</span>
+                                    ))}</td>
+                                    <td className='p-4 align-middle'>
+                                        <div className='flex items-center justify-end gap-3'>
+                                            <CgEditExposure size={18} className='text-blue-500 cursor-pointer hover:text-blue-300' onClick={() => handleSubCategory(cat)} />
+                                            <MdEdit size={18} className='text-blue-500 cursor-pointer hover:text-blue-300' onClick={() => handleUpdateCategory(cat)} />
+                                            <CgTrashEmpty size={18} className='text-red-500 cursor-pointer hover:text-red-300' onClick={() => handleCategoryDelete(cat._id)} />
+                                        </div>
+                                    </td>
+                                </tr>
+                            ))
+                        ) : (
+                            <tr>
+                                <td colSpan={6} className='text-center py-6 text-gray-500'>
+                                    No category created yet
+                                </td>
+                            </tr>
+                        )}
+                    </tbody>
+                </table>
             </div>
 
             {/* Models */}
 
-            <div className={`category-model fixed top-0 left-0 bg-[#9393931c] w-full min-h-full px-2 py-10 z-100 transition-all duration-200 ease-linear ${openModel ? 'category-model-open opacity-100' : 'category-model-close opacity-50'}`}>
-                <div className={`category-model-box w-full max-w-[500px] bg-white p-5 rounded-[12px] transition-all duration-200 ease-linear ${openModel ? 'scale-100' : 'scale-75'}`}>
+            {/* Create Category  */}
+            <div className={`category-model fixed top-0 left-0 bg-[#9393931c] w-full min-h-full px-2 py-10 z-100 transition-all duration-200 ease-linear ${openModel ? 'category-model-open opacity-100' : 'category-model-close opacity-0'}`}>
+                <div className={`category-model-box w-full max-w-[500px] bg-white p-5 rounded-[12px] transition-all duration-200 ease-linear ${openModel ? 'scale-100' : 'scale-90'}`}>
                     <p className='font-bold !text-[#333] flex gap-1 items-center'><BiInfoCircle /> Add New Category</p>
                     <p className='mb-4'>Create a new category for organizing images</p>
-                    <input type="text" name="category" id="category" value={createCategoryInput} placeholder='Enter category to add' className='w-full px-3 py-2 bg-white rounded-[12px]' onChange={(e) => setCreateCategoryInput(e.target.value)} />
+                    <div>
+                        <label className='mb-1 font-semibold'>Name</label>
+                        <input type="text" name="category" id="category" value={createCategoryInput} placeholder='Enter category name' className='w-full px-3 py-2 bg-white rounded-[12px]' onChange={(e) => setCreateCategoryInput(e.target.value)} />
+                    </div>
+                    <div>
+                        <label className='mb-1 font-semibold mt-4'>Select Image</label>
+                        <input type="file" name="image" id="image" placeholder='Select image for category' className='w-full px-3 py-2 bg-white cursor-pointer rounded-[12px]' onChange={handleFileChange} />
+                    </div>
+                    {
+                        preview &&
+                        <img src={preview} alt="preview" className='w-[150px] object-contain mt-4' />
+                    }
                     <div className='flex gap-3 justify-end mt-6'>
                         <button className='px-[20px] py-[8px] bg-[#ddd] rounded-[10px] transition-all duration-200 ease-in hover:bg-[#d2d2d2]'
                             onClick={() => setOpenModel(false)}
@@ -262,25 +330,56 @@ export default function Categories() {
                 </div>
             </div>
 
-            <div className={`category-model fixed top-0 left-0 bg-[#9393931c] w-full min-h-full px-2 py-10 z-100 transition-all duration-200 ease-linear ${openCatModel ? 'category-model-open opacity-100' : 'category-model-close opacity-50'}`}>
-                <div className={`category-model-box w-full max-w-[500px] bg-white p-5 rounded-[12px] transition-all duration-200 ease-linear ${openCatModel ? 'scale-100' : 'scale-75'}`}>
+            {/* Update Category */}
+            <div className={`category-model fixed top-0 left-0 bg-[#9393931c] w-full min-h-full px-2 py-10 z-100 transition-all duration-200 ease-linear ${openCatModel ? 'category-model-open opacity-100' : 'category-model-close opacity-0'}`}>
+                <div className={`category-model-box w-full max-w-[500px] bg-white p-5 rounded-[12px] transition-all duration-200 ease-linear ${openCatModel ? 'scale-100' : 'scale-90'}`}>
                     <p className='font-bold !text-[#333] mb-5 flex gap-1 items-center'><BiInfoCircle /> Update Category</p>
-                    <input type="text" name="category" id="category" value={updatingCategory?.category} placeholder='Enter category to update' className='w-full px-3 py-2 bg-white rounded-[12px]' onChange={(e) => setUpdatingCategory({ ...updatingCategory, category: e.target.value.toLowerCase() })} />
+                    <div>
+                        <label className='mb-1 font-semibold'>Name</label>
+                        <input type="text" name="category" id="category" value={updatingCategory?.category} placeholder='Enter category to update' className='w-full px-3 py-2 bg-white rounded-[12px]' onChange={(e) => setUpdatingCategory({ ...updatingCategory, category: e.target.value.toLowerCase() })} />
+                    </div>
+                    <div>
+                        <label className='mb-1 font-semibold mt-4'>Image</label>
+                        <input type="file" name="image" id="image" placeholder='Select image for category' className='w-full px-3 py-2 bg-white cursor-pointer rounded-[12px]' onChange={handleFileChange} />
+                        {updatingCategory?.imageURL && !preview && (
+                            <div className="relative mb-3">
+                                <img
+                                    src={`${import.meta.env.VITE_ASURA_SUBDOMAIN}${updatingCategory.imageURL}`}
+                                    alt="Current category"
+                                    className='w-[150px] h-[150px] object-contain border rounded-lg'
+                                />
+                            </div>
+                        )}
+                        {preview && (
+                            <div className="relative mt-3">
+                                <img
+                                    src={preview}
+                                    alt="Preview"
+                                    className='w-[150px] h-[150px] object-contain border rounded-lg'
+                                />
+                            </div>
+                        )}
+                    </div>
                     <div className='flex gap-3 justify-end mt-6'>
                         <button className='px-[20px] py-[8px] bg-[#ddd] rounded-[10px] transition-all duration-200 ease-in hover:bg-[#d2d2d2]'
-                            onClick={() => setOpenCatModel(false)}
+                            onClick={() => {
+                                setOpenCatModel(false);
+                                setPreview(null);
+                                setSelectedFile(null);
+                            }}
                         >
                             Cancel
                         </button>
-                        <button className='px-[20px] py-[8px] text-[var(--x-light)] bg-[var(--dark)] rounded-[10px] flex gap-2 items-center transition-all duration-200 ease-in hover:bg-[var(--md-dark)]' onClick={handleUpdateCategoryFunction}>
-                            Update category
+                        <button className='px-[20px] py-[8px] text-[var(--x-light)] bg-[var(--dark)] rounded-[10px] flex gap-2 items-center transition-all duration-200 ease-in hover:bg-[var(--md-dark)]' disabled={loading} onClick={handleUpdateCategoryFunction}>
+                            {loading ? 'Updating...' : 'Update category'}
                         </button>
                     </div>
                 </div>
             </div>
 
-            <div className={`category-model fixed top-0 left-0 bg-[#9393931c] w-full min-h-[100vh] px-2 py-10 z-100 transition-all duration-200 ease-linear ${openSubCatModel ? 'category-model-open opacity-100' : 'category-model-close opacity-50'}`}>
-                <div className={`category-model-box w-full max-w-[500px] max-h-[90vh] overflow-y-scroll bg-white p-5 rounded-[12px] transition-all duration-200 ease-linear ${openSubCatModel ? 'scale-100' : 'scale-75'}`}>
+            {/* Add Subcategory */}
+            <div className={`category-model fixed top-0 left-0 bg-[#9393931c] w-full min-h-[100vh] px-2 py-10 z-100 transition-all duration-200 ease-linear ${openSubCatModel ? 'category-model-open opacity-100' : 'category-model-close opacity-0'}`}>
+                <div className={`category-model-box w-full max-w-[500px] max-h-[90vh] overflow-y-scroll bg-white p-5 rounded-[12px] transition-all duration-200 ease-linear ${openSubCatModel ? 'scale-100' : 'scale-90'}`}>
                     <p className='font-bold !text-[#333] mb-5 flex gap-1 items-center'><BiInfoCircle /> Add New Subcategory</p>
                     <input type="text" name="category" id="category" value={subcategory} placeholder='Enter subcategory to add' className='w-full px-3 py-2 bg-white rounded-[12px]' onChange={(e) => setSubcategory(e.target.value.toLowerCase())} />
                     <div className='flex gap-3 justify-end mt-6'>
